Add 'resolve' method
